requirements.txt
fastapi
uvicorn
python-dotenv
openai
parselmouth
numpy